Implement Professional Reports PDF export using jsPDF, add Viewer and Entry Staff role guards on database endpoints and sidebar layouts

+++ expense.php
@@ -3,6 +3,21 @@
 require_once 'config.php';
 require_login();
 
+// Role guards: Viewer cannot access this page. Staff cannot edit or delete.
+if ($_SESSION['user_role'] === 'viewer') {
+    set_flash_message('error', 'Viewer accounts are restricted to reports and read-only actions.');
+    header("Location: reports.php");
+    exit;
+}
+
+if ($_SESSION['user_role'] === 'staff') {
+    if (isset($_GET['edit']) || isset($_GET['delete'])) {
+        set_flash_message('error', 'Access denied. Entry staff members cannot edit or delete transactions.');
+        header("Location: expense.php");
+        exit;
+    }
+}
+
 $active_page = 'expense';
 $error = '';
 $success = '';
@@ -100,6 +115,9 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
                 $current_balance = (float)($account['current_balance'] ?? 0);
                 
                 if (isset($_POST['action']) && $_POST['action'] === 'update') {
+                    if ($_SESSION['user_role'] === 'staff') {
+                        throw new Exception("Access denied. Entry staff members cannot modify transactions.");
+                    }
                     // UPDATE MODE
                     $exp_id = (int)$_POST['expense_id'];
                     
@@ -289,12 +307,14 @@ $categories = ['Office Rent', 'Utilities (Electricity/Water)', 'Internet / Subsc
                                                 <a href="invoice.php?type=expense&id=<?= $exp['id'] ?>" class="btn-icon btn-print" title="Print Invoice" target="_blank">
                                                     <i class="fa-solid fa-print"></i>
                                                 </a>
-                                                <a href="?edit=<?= $exp['id'] ?>" class="btn-icon btn-edit" title="Edit">
-                                                    <i class="fa-solid fa-pen"></i>
-                                                </a>
-                                                <a href="?delete=<?= $exp['id'] ?>" class="btn-icon btn-delete" title="Delete" onclick="return confirm('Are you sure you want to delete this expense? The amount will be refunded back to the bank account.');">
-                                                    <i class="fa-solid fa-trash"></i>
-                                                </a>
+                                                <?php if ($_SESSION['user_role'] !== 'staff'): ?>
+                                                    <a href="?edit=<?= $exp['id'] ?>" class="btn-icon btn-edit" title="Edit">
+                                                        <i class="fa-solid fa-pen"></i>
+                                                    </a>
+                                                    <a href="?delete=<?= $exp['id'] ?>" class="btn-icon btn-delete" title="Delete" onclick="return confirm('Are you sure you want to delete this expense? The amount will be refunded back to the bank account.');">
+                                                        <i class="fa-solid fa-trash"></i>
+                                                    </a>
+                                                <?php endif; ?>
                                             </td>
                                         </tr>
                                     <?php endforeach; ?>
